requirements.txt
asgiref==3.2.10
astroid==2.4.2
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Django==3.0.8
django-jsonfield==1.4.0
django-materializecss-form==1.1.17
django-model-utils==4.0.0
django-nested-admin==3.3.2
django-redis==4.12.1
django-redis-cache==2.1.1
django-rest-auth==0.9.5
django-tabular-permissions==2.6
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
docx==0.2.4
docx2txt==0.8
docxtpl==0.10.0
google-api-core==1.21.0
google-auth==1.18.0
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
isort==4.3.21
Jinja2==2.11.2
lazy-object-proxy==1.4.3
lxml==4.5.1
MarkupSafe==1.1.1
mccabe==0.6.1
Pillow==7.2.0
pkg-resources==0.0.0
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycode==0.1.2
PyJWT==1.7.1
pylint==2.5.3
python-docx==0.8.10
python-monkey-business==1.0.0
pytz==2020.1
redis==3.5.3
requests==2.24.0
rsa==4.6
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
wrapt==1.12.1
